fix(smart-sessions): pack time-frame initData as uint48 || uint48 (12 bytes)
Deployed TimeFramePolicy slices `initData[0:12]` and unpacks it as
`uint96 → uint48 validUntil || uint48 validAfter` (high 48 bits || low 48 bits).
The previous `encodePacked(['uint128','uint128'], …)` wrote 32 bytes; the
contract took the high 12 bytes of the first `uint128` — all zero for any
plausible timestamp — and stored `TimeFrameConfig(0)`. Init silently
succeeded, then every later `_checkTimeFrame` reverted with
`PolicyNotInitialized`.

Encode as `uint48 || uint48` so the 12-byte layout matches the deployed
contract.

+++ src/modules/validators/smart-sessions.ts
@@ -1155,14 +1155,15 @@ function getPolicyData(
       }
     }
     case 'time-frame': {
-      // Deployed TimeFramePolicy reads `bytes16 validUntil || bytes16 validAfter` (32 bytes).
+      // Deployed TimeFramePolicy slices initData[0:12] and unpacks as
+      // uint48 validUntil || uint48 validAfter (high 48 bits || low 48 bits).
       return {
         policy: addresses.timeFrame,
         initData: encodePacked(
-          ['uint128', 'uint128'],
+          ['uint48', 'uint48'],
           [
-            BigInt(Math.floor(policy.validUntil / 1000)),
-            BigInt(Math.floor(policy.validAfter / 1000)),
+            Math.floor(policy.validUntil / 1000),
+            Math.floor(policy.validAfter / 1000),
           ],
         ),
       }
